refactor: Move helpers from views.py to utils.py
